+++ config/constant.ts
@@ -24,7 +24,7 @@ export const MOCK_API_TARGET_URL = 'http://localhost:3000';
 export const ICONFONTURL = '//at.alicdn.com/t/font_3004192_9jmc1z9neiw.js'; // 去色版
 
 // 包依赖分析
-export const ANALYSIS = false;
+export const ANALYSIS = true;
 
 // 是否支持Md渲染
 export const MARKDOWN = true;
